fix: split up the executions and execution for pipe pages
