+++ packages/frontend/src/pages/Executions/ExecutionsForFlowPage.tsx
@@ -1,5 +1,7 @@
-import type { IExecution, IFlow } from '@plumber/types'
+import type { IExecution } from '@plumber/types'
 
+import { useEffect } from 'react'
+import { useParams } from 'react-router-dom'
 import { useQuery } from '@apollo/client'
 import { Center, Flex } from '@chakra-ui/react'
 import { Pagination } from '@opengovsg/design-system-react'
@@ -10,9 +12,10 @@ import NoResultFound from '@/components/NoResultFound'
 import PageTitle from '@/components/PageTitle'
 import PrimarySpinner from '@/components/PrimarySpinner'
 import { GET_EXECUTIONS } from '@/graphql/queries/get-executions'
+import { GET_FLOW } from '@/graphql/queries/get-flow'
 import { usePaginationAndFilter } from '@/hooks/usePaginationAndFilter'
 
-import StatusInput from './StatusInput'
+import StatusInput from './components/StatusInput'
 
 const RESULTS_PER_PAGE = 10
 
@@ -26,6 +29,10 @@ const getLimitAndOffset = (page: number) => ({
   limit: RESULTS_PER_PAGE,
   offset: (page - 1) * RESULTS_PER_PAGE,
 })
+
+type ExecutionsForFlowParams = {
+  flowId: string
+}
 
 function ExecutionsList({
   executions,
@@ -66,30 +73,65 @@ function ExecutionsList({
   )
 }
 
-export default function ExecutionList({ flow }: { flow: IFlow }) {
+export default function ExecutionsForFlowPage() {
+  const { flowId } = useParams() as ExecutionsForFlowParams
+
   const { page, setSearchParams, status, isSearching } =
     usePaginationAndFilter()
-  const { id: flowId, name: flowName } = flow
-  const { data: executionsData, loading } = useQuery(GET_EXECUTIONS, {
-    variables: {
-      ...getLimitAndOffset(page),
-      status,
-      flowId,
-    },
-    fetchPolicy: 'cache-and-network',
+
+  const { data: flowData, loading: isFlowLoading } = useQuery(GET_FLOW, {
+    variables: { id: flowId },
     skip: !flowId,
   })
+
+  const { data: executionsData, loading: isExecutionsLoading } = useQuery(
+    GET_EXECUTIONS,
+    {
+      variables: {
+        ...getLimitAndOffset(page),
+        status,
+        flowId,
+      },
+      skip: !flowId,
+    },
+  )
 
   const { pageInfo, edges } = executionsData?.getExecutions || {}
   const executions: IExecution[] =
     edges?.map(({ node }: { node: IExecution }) => {
       return {
         ...node,
-        flow,
+        flow: flowData?.getFlow,
       }
     }) ?? []
   const totalCount: number = pageInfo?.totalCount ?? 0
-  const hasPagination = !loading && totalCount > RESULTS_PER_PAGE
+  const isLoading = isExecutionsLoading || isFlowLoading
+  const hasPagination = !isLoading && totalCount > RESULTS_PER_PAGE
+
+  // ensure invalid pages won't be accessed even after deleting flows
+  const lastPage = Math.ceil(totalCount / RESULTS_PER_PAGE)
+
+  useEffect(() => {
+    // Defer the search params update till after the initial render
+    if (lastPage !== 0 && page > lastPage) {
+      setSearchParams({ page: lastPage })
+    }
+  }, [lastPage, page, setSearchParams])
+
+  if (isLoading) {
+    return (
+      <Center mt={8}>
+        <PrimarySpinner fontSize="4xl" />
+      </Center>
+    )
+  }
+
+  if (!flowData?.getFlow) {
+    return <NoResultFound description="Flow not found" />
+  }
+
+  const { name: flowName } = flowData.getFlow
+
   return (
     <Container py={9}>
       <PageTitle
@@ -105,7 +147,7 @@ export default function ExecutionList({ flow }: { flow: IFlow }) {
       />
       <ExecutionsList
         executions={executions}
-        isLoading={loading}
+        isLoading={isLoading}
         isSearching={isSearching}
       />
       {hasPagination && (
