+++ packages/frontend/src/routes.tsx
@@ -9,6 +9,7 @@ import Applications from '@/pages/Applications'
 import EditorRoutes from '@/pages/Editor/routes'
 import Execution from '@/pages/Execution'
 import Executions from '@/pages/Executions'
+import ExecutionsForFlowPage from '@/pages/Executions/ExecutionsForFlowPage'
 import Flow from '@/pages/Flow'
 import Flows from '@/pages/Flows'
 import Login from '@/pages/Login'
@@ -39,6 +40,15 @@ export default createRoutesFromElements(
       element={
         <Layout>
           <Executions />
+        </Layout>
+      }
+    />
+
+    <Route
+      path={URLS.EXECUTIONS_FOR_FLOW_PATTERN}
+      element={
+        <Layout>
+          <ExecutionsForFlowPage />
         </Layout>
       }
     />
